git-svn-id: https://svn.osdn.net/svnroot/cria@368 1d3d8b80-664f-4c7c-9890-4cefb2de5291

+++ src/Cria/source/Reference.h
@@ -67,7 +67,7 @@ struct ReferenceLiteralTag
     ReferenceLiteralType type;
     union {
         String string;
-        int integer;
+        long integer;
         Boolean boolean;
     } of;
 };
